ResponseLogFormatter follows same code pattern RequestLogFormatter
