+++ src/main/scala/nl.knaw.dans.easy.deposit/logging/PlainAuthHeaders.scala
@@ -18,9 +18,9 @@ package nl.knaw.dans.easy.deposit.logging
 import javax.servlet.http.HttpServletResponse
 
 trait PlainAuthHeaders extends ResponseLogFormatter {
-  override protected def maskAuthHeaders(response: HttpServletResponse): Array[String] = {
+  override protected def maskAuthHeaders(response: HttpServletResponse): Map[String, String] = {
     response.getHeaderNames.toArray().map {
-      case name: String => s"$name -> ${response.getHeader(name)}"
-    }
+      case name: String => name -> response.getHeader(name)
+    }.toMap
   }
 }
